pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentcontract"
version = "0.1.0"
description = "Behavioral contracts for AI agents — Python reference implementation"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Mauro Moro", email = "mauromoro@msn.com" }]
keywords = ["ai", "agents", "contracts", "llm", "safety", "compliance", "validation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.40"]
schema = ["jsonschema>=4.0"]
all = ["anthropic>=0.40", "jsonschema>=4.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "anthropic>=0.40",
    "jsonschema>=4.0",
]

[project.scripts]
agentcontract = "agentcontract.cli:cli"

[project.urls]
Homepage = "https://github.com/agentcontract/spec"
Repository = "https://github.com/agentcontract/agentcontract-py"
Issues = "https://github.com/agentcontract/agentcontract-py/issues"
Specification = "https://github.com/agentcontract/spec/blob/main/SPEC.md"

[tool.hatch.build.targets.wheel]
packages = ["src/agentcontract"]

[tool.pytest.ini_options]
testpaths = ["tests"]
